simplify jsx content
update file name
move InitZoom function to another file
move shared html content to another file
